+++ gemm_fp/gemm_cpu_fp.cpp
@@ -23,15 +23,15 @@ void cpu_gemm_nn(int TA, int TB, int M, int N, int K, float ALPHA,
 }
 
 void fixed_cpu_gemm_nn(int TA, int TB, int M, int N, int K, float ALPHA,
-        long long int *A, int lda,
-        long long int *B, int ldb,
+        long int *A, int lda,
+        long int *B, int ldb,
         float BETA,
-        long long int *C, int ldc)
+        long int *C, int ldc)
 {
     int i,j,k;
     for(i = 0; i < M; ++i){
         for(k = 0; k < K; ++k){
-            PUT_IN_REGISTER long long int A_PART = ALPHA * A[i * lda + k];
+            PUT_IN_REGISTER long int A_PART = ALPHA * A[i * lda + k];
             for(j = 0; j < N; ++j){
                 C[i*ldc+j] += A_PART*B[k*ldb+j];
             }
